LayoutTests/imported/w3c/web-platform-tests/cssom-view/elementFromPoint-002-expected.txt


FAIL Checking whether dynamic changes to visibility interact correctly with
  table anonymous boxes assert_equals: Should hit the overlay first. expected Element node <div id="overlay"><div></div></div> but got Element node <div id="target">Some text</div>
Some text
